+++ Unity/Assets/Scripts/CUBES/BlockData.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 using System.Collections;
 
 public class BlockData : MonoBehaviour {
-	public int id;
+	public sbyte id;
 	public int metadata;
 	public int reference;
 }
